Created utils for api request.
- Added function to create the body required for every request.
- Added function to get the request path.
